Allow all of Sequel::DataSet#join parameters
Dataset#join takes four parameters, though the last is not often used.

The actual signature of the method is:
  `def join_table(type, table, expr=nil, options=OPTS, &block)`

where expr is the basic set of `on` parameters for the join
conditions, and OPTS is a set of table modifiers.

For docs, see https://github.com/jeremyevans/sequel/blob/master/lib/sequel/dataset/query.rb#L470-L528