Make it possible to build associations on top of custom relations
This allows associations to receive the target relation as an argument
and it'll use it instead of reaching for the canonical target relation.

auto-combine was changed so that it always pass `self` as the target
relation, this way we can do things like add other joins, additional
columns etc. when combining via associations.

The biggest benefit is that wrap+combine can now be used together