Merge association schema with view schema (refs #123)
Additional view for associations means that we want to extend the
relation created by an association, so it didn't make sense to project
view relation with the association schema, as it would override
attributes provided by views. That's why we want to merge them, and make
sure we have unique attributes in the resulting schema.