Add a warning when inference fails
This will probably cause warnings with Rails because it always tries to
load ROM when loading app environment. So ie in migration tasks you're
gonna see this. This can be improved by some configuration where we can
disable this warning for particular usecases, but this is Rails-specific
so it should go into rom-rails.

Refs #137