Remove constraint from Serial
Because all relations have schemas now, this turned out to be a
regression as in cases where people pass `nil` as PK, Sequel will simply
ignore it and let the db generate a serial PK value. After adding
schemas to all relations that constraint we have here made commands raise
exceptions when `id` was passed in with `nil` value.

Given that this was a regression, and questionable value of this
constraint, it's better to remove it.