Return a relation with empty schema from `Relation#read`
Since we can't tell what the attributes are, it's better to return an
empty schema. This way repo can generate open structs for such
relations. You may notice that `#columns` still return attribute names
BUT it actually executes a query to see what the columns are, and we
definitely don't want to do this.

The best approach with `#read` is to use it along with `view` DSL where
you can define schema upfront, ie:

  view(:names) do
    schema { project(:name) }
    relation { read('select name from users') }
  end