Wrap JSON primitives before passing them to Sequel (fix #327)
This adds support for writing primitive values to JSON(B) columns. Sequel has recently added support for it. On the rom-sql's side we only need to change the wrapper.

Other values are interpreted as JSONOp as they did before. We could revise this behavior in the next major version, it was added for compatibility with Hanami < 1.0 but I'm not sure if we still need it.