Don't remove order in subqueries
Although removing ORDER BY part from a subquery
may seem an innocent change, it really isn't
and it leads to bugs.

Consider the following query:

```
SELECT *
FROM tasks
WHERE user_id IN (SELECT id FROM users ORDER BY name LIMIT 1)
```

Here, we list all the task of the user whose name is
alphabetically first.

If we remove the order, however, this query becomes
non-deterministic and tasks of a random user are returned
(but typically, it's the user with the smallest id).

So, let's not be overly smart and not do `unordered`
for the subqueries. Let's be explicit.