Document `:as` option for `schema`
If we are migrating from a legacy database to a new one, we might want
to use a different name to refer to database tables in the legacy
database even though the new database has the same table names.

The `:as` option allows us to do this but because we couldn't find any
documentation for how to use it in `rom-sql`, we have added an example.

Co-authored-by: Steven Chudley <steven.chudley@gmail.com>